conversation/requirements.txt
openai==1.64.0
langchain==0.3.19
langchain_community==0.3.18
SpeechRecognition==3.8.1
pydub==0.25.1

# ffmpeg needed for whisper
openai-whisper

transformers==4.49.0
# torch==2.6.0
dotenv
langchain_openai
pymongo
numpy
pyAudioAnalysis
scipy
matplotlib
gtts
langdetect
gradio
loguru